# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZenFM API
  version: 0.1.0
  description: |
    The local, single-owner ZenFM HTTP API. Browser requests authenticate with
    an opaque `zenfm_session` cookie and send the session-bound value returned
    as `csrfToken` in `X-ZenFM-CSRF` for every mutation. Personal API tokens use
    `Authorization: Bearer`; their restricted scope excludes owner security,
    token management, update, and lifecycle operations.
servers:
  - url: /
paths:
...
  /api/v1/files:
    get:
      operationId: listFiles
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
        - $ref: '#/components/parameters/Hidden'
      responses:
        '200':
          description: One directory. Special entries are metadata-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
        - $ref: '#/components/parameters/Path'
        - in: query
          name: recursive
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Target deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/files/directory:
    post:
      operationId: createDirectory
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [path]
              properties:
                path:
                  $ref: '#/components/schemas/Path'
      responses:
        '201':
          description: Directory created.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/files/content:
    get:
      operationId: getFileContent
      description: Returns exact, bounded UTF-8 source for the text editor; unlike preview output, HTML is not transformed.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
      responses:
        '200':
          description: Exact editable text, limited to 4 MiB.
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/Unsupported'
    put:
      operationId: putFile
      description: Atomically creates or explicitly replaces a regular file. Symlinks and special files are rejected.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
        - $ref: '#/components/parameters/Path'
        - in: header
          name: If-None-Match
          schema:
            type: string
          description: Use `*` to reject overwrite conflicts.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: File created.
        '204':
          description: File replaced.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/TooLarge'
  /api/v1/files/move:
    post:
      operationId: moveFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '204':
          description: Target moved or renamed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/files/copy:
    post:
      operationId: copyFile
      description: Recursion skips symlinks, pseudo-filesystems, and non-regular entries.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '204':
          description: Target copied.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/files/checksum:
    get:
      operationId: checksumFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
        - in: query
          name: algorithm
          schema:
            type: string
            enum: [sha256, sha512]
            default: sha256
      responses:
        '200':
          description: Digest of the regular file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checksum'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/search:
    get:
      operationId: searchFiles
      description: Recursively searches regular files/directories, skipping pseudo-filesystems and special entries.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
        - $ref: '#/components/parameters/Hidden'
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 250
      responses:
        '200':
          description: Bounded search result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '408':
          $ref: '#/components/responses/TimedOut'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zenfm_session
    personalToken:
      type: http
      scheme: bearer
      bearerFormat: opaque ZenFM personal token
  parameters:
...
    CsrfOptionalForBearer:
      in: header
      name: X-ZenFM-CSRF
      required: false
      schema:
        type: string
      description: Required for cookie-authenticated mutations; ignored for bearer authentication.
    Path:
      in: query
      name: path
      required: true
      schema:
        $ref: '#/components/schemas/Path'
    Hidden:
      in: query
      name: hidden
      required: false
      description: Overrides the stored hidden-file preference for this listing or search.
      schema:
        type: boolean
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, or revoked credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Forbidden:
      description: Authenticated principal lacks access, setup is incomplete, CSRF failed, or origin was rejected.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Target or upload state conflicts with the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooLarge:
      description: A configured body, upload, preview, or archive bound was exceeded.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unsupported:
      description: The target cannot be safely previewed or opened as regular content.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TimedOut:
      description: Server-side operation deadline elapsed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Path:
      type: string
      minLength: 1
      maxLength: 4096
      pattern: '^/'
      examples: [/Books]
...
    FileEntry:
      type: object
      additionalProperties: false
      required: [name, path, type, size, modifiedAt]
      properties:
        name:
          type: string
        path:
          $ref: '#/components/schemas/Path'
        type:
          type: string
          enum: [file, directory, symlink, special]
        size:
          type: integer
          minimum: 0
        modifiedAt:
          type: string
          format: date-time
        mimeType:
          type: string
        hidden:
          type: boolean
        writable:
          type: boolean
    FileListing:
      type: object
      additionalProperties: false
      required: [path, entries, advancedMode]
      properties:
        path:
          $ref: '#/components/schemas/Path'
        entries:
          type: array
          maxItems: 100000
          items:
            $ref: '#/components/schemas/FileEntry'
        advancedMode:
          type: boolean
        disk:
          $ref: '#/components/schemas/Usage'
    SearchResult:
      type: object
      additionalProperties: false
      required: [entries, truncated]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/FileEntry'
        truncated:
          type: boolean
    Transfer:
      type: object
      additionalProperties: false
      required: [source, destination]
      properties:
        source:
          $ref: '#/components/schemas/Path'
        destination:
          $ref: '#/components/schemas/Path'
    Checksum:
      type: object
      additionalProperties: false
      required: [algorithm, value]
      properties:
        algorithm:
          type: string
          enum: [sha256, sha512]
        value:
          type: string
          pattern: '^[0-9a-f]+$'
...
    Usage:
      type: object
      additionalProperties: false
      required: [used, total]
      properties:
        used:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    Problem:
      type: object
      additionalProperties: true
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference